pyproject.toml
[project]
name = "SaBBLium"
dynamic = ["version"]
authors = [
    { name = "Ludovic Denoyer" },
    { name = "Alfredo de la Fuente"},
    { name = "Song Duong"},
    { name = "Jean-Baptiste Gaya"},
    { name = "Pierre-Alexandre Kamienny"},
    { name = "Daniel H. Thompson"},
    { name = "Olivier Sigaud", email = "olivier.sigaud@isir.upmc.fr" },
    { name = "Mathis Koroglu", email = "mathis.koroglu@etu.sorbonne-universite.fr" }
]
maintainers = [
    { name = "Mathis Koroglu", email = "mathis.koroglu@etu.sorbonne-universite.fr" }
]

description = "SaBBLium is a Flexible and Simple Library for Learning Sequential Agents"
keywords = ["sequential learning", "reinforcement learning"]

readme = "README.md"
license = { file = "LICENSE" }

requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    "torchvision",
    "gymnasium",
]

[project.optional-dependencies]
plot = ["matplotlib", "seaborn"]
format = ["black", "isort"]

[project.urls]
repository = "https://github.com/Arlaz/SaBBLium"

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sabblium*"]
namespaces = true

[tool.setuptools_scm]
write_to = "sabblium/_version.py"

[tool.yapf]
allow_multiline_lambdas = true
allow_multiline_dictionary_keys = true
arithmetic_precedence_indication = true
based_on_style = "pep8"
column_limit = 120
coalesce_brackets = false
indent_width = 4
indent_closing_brackets = false
spaces_before_comment = 1
split_before_logical_operator = true
spaces_around_power_operator = true
spaces_around_default_or_named_assign = false
spaces_around_dict_delimiters = false
spaces_around_list_delimiters = false
spaces_around_tuple_delimiters = false
spaces_around_subscript_colon = false
split_before_bitwise_operator = true
split_before_dot = true
split_complex_comprehension = true

# ... other project metadata fields as specified in:
#     https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
